Refactor chunk_typename
Simple move
